requirements.txt
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-environ==0.12.0
django-jazzmin==3.0.1
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
drf-yasg==1.21.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.19
pillow==12.0.0
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
